Title: 최솟값 3개, Time: 1741ms, Memory: 46MB, Status: Passed - CodeTree

+++ 241023/최솟값 3개/top-3-smallest-number.java
@@ -23,7 +23,7 @@ public class Main {
                 pq.add(y);
                 pq.add(z);
 
-                System.out.println(x * y * z);
+                System.out.println((long) x * y * z);
             } else {
                 System.out.println(-1);
             }
